Creating game updates

+++ app/[boardId]/[mapId]/game/play-header.tsx
@@ -1,5 +1,6 @@
 'use client'
 import { useEffect, useState } from "react";
+import Swal from 'sweetalert2'
 import { deleteGameState } from "@/lib/store/gameState";
 import gameStateSyncService from "./game-state-sync-service";
 
@@ -12,16 +13,25 @@ export default function PlayHeader(
   useEffect(() => gameStateSyncService.subscribe(boardId, mapId, setGameState), [boardId, mapId]);
 
   const deleteGameAction = async () => {
-    await deleteGameState(boardId, mapId);
+    const result = await Swal.fire({
+      title: 'Reset game?',
+      icon: 'warning',
+      text: "This will delete your current game and start a new game. This action cannot be undone!",
+      showCancelButton: true,
+      confirmButtonColor: 'var(--color-error)',
+      confirmButtonText: 'Delete!'
+    })
+
+    if (result.isConfirmed) {
+      await deleteGameState(boardId, mapId);
+    }
   };
 
   if (!gameState) {
-    return null;
+    return <h3>Welcome to TableRunner</h3>;
   }
 
   return (
-    <form action={deleteGameAction}>
-      <button type="submit">X</button>
-    </form>
+    <button type="submit" onClick={deleteGameAction}>X</button>
   );
 }
